refactor: split logic and cli into separate files
